Refactor and clean code
