+++ CodeAnalyzerRules.ruleset
@@ -93,8 +93,7 @@
     <Rule Id="IDE0003" Action="None" />
   </Rules>
   <Rules AnalyzerId="Roslynator.CSharp.Analyzers" RuleNamespace="Roslynator.CSharp.Analyzers">
-    <Rule Id="RCS1080" Action="Hidden" />
-    <Rule Id="RCS1090" Action="Hidden" />
+    <Rule Id="RCS1090" Action="None" />
   </Rules>
   <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers">
     <Rule Id="SA0001" Action="Hidden" />
