+++ Esbc.sln
@@ -5,15 +5,11 @@ VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{E8B74D0A-A998-485B-9948-0B7837919310}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Consumer", "src\EsbcConsumer\EsbcConsumer.csproj", "{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Producer", "src\EsbcProducer\EsbcProducer.csproj", "{DA0B1BA3-F7A9-4BEE-8F8F-6D1F6CD4A531}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "__tests__", "__tests__", "{0E5FE3E9-74CE-41D4-96ED-6074A8537F43}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Consumer", "__tests__\EsbcConsumer\EsbcConsumer.csproj", "{5327E402-F93C-4EB0-8C96-72343B448C08}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Producer", "__tests__\EsbcProducer\EsbcProducer.csproj", "{45CD6946-0187-493B-AF55-39209D37C104}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProducerTest", "__tests__\EsbcProducerTest\EsbcProducerTest.csproj", "{45CD6946-0187-493B-AF55-39209D37C104}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,18 +24,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Debug|x64.Build.0 = Debug|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Debug|x86.Build.0 = Debug|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Release|x64.ActiveCfg = Release|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Release|x64.Build.0 = Release|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Release|x86.ActiveCfg = Release|Any CPU
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1}.Release|x86.Build.0 = Release|Any CPU
 		{DA0B1BA3-F7A9-4BEE-8F8F-6D1F6CD4A531}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DA0B1BA3-F7A9-4BEE-8F8F-6D1F6CD4A531}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DA0B1BA3-F7A9-4BEE-8F8F-6D1F6CD4A531}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -52,18 +36,6 @@ Global
 		{DA0B1BA3-F7A9-4BEE-8F8F-6D1F6CD4A531}.Release|x64.Build.0 = Release|Any CPU
 		{DA0B1BA3-F7A9-4BEE-8F8F-6D1F6CD4A531}.Release|x86.ActiveCfg = Release|Any CPU
 		{DA0B1BA3-F7A9-4BEE-8F8F-6D1F6CD4A531}.Release|x86.Build.0 = Release|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Debug|x64.Build.0 = Debug|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Debug|x86.Build.0 = Debug|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Release|x64.ActiveCfg = Release|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Release|x64.Build.0 = Release|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Release|x86.ActiveCfg = Release|Any CPU
-		{5327E402-F93C-4EB0-8C96-72343B448C08}.Release|x86.Build.0 = Release|Any CPU
 		{45CD6946-0187-493B-AF55-39209D37C104}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{45CD6946-0187-493B-AF55-39209D37C104}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{45CD6946-0187-493B-AF55-39209D37C104}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -78,9 +50,7 @@ Global
 		{45CD6946-0187-493B-AF55-39209D37C104}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{3B4A0CBC-104B-4ABB-BF78-DCEB0FF8F4F1} = {E8B74D0A-A998-485B-9948-0B7837919310}
 		{DA0B1BA3-F7A9-4BEE-8F8F-6D1F6CD4A531} = {E8B74D0A-A998-485B-9948-0B7837919310}
-		{5327E402-F93C-4EB0-8C96-72343B448C08} = {0E5FE3E9-74CE-41D4-96ED-6074A8537F43}
 		{45CD6946-0187-493B-AF55-39209D37C104} = {0E5FE3E9-74CE-41D4-96ED-6074A8537F43}
 	EndGlobalSection
 EndGlobal
